pyproject.toml
[project]
name = "playwright-shell"
version = "0.1.0"
description = "Modular browser automation shell built on Playwright, optional PyAutoGUI, and uv."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.53.0",
    "pyautogui>=0.9.54",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.10.0",
    "pyyaml>=6.0.2",
    "typer>=0.16.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "ruff>=0.12.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project.scripts]
playwright-shell = "playwright_shell.cli:app"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
